refactor : Properties 클래스를 global.properties 패키지로 이동
