+++ src/main/java/_1danhebojo/coalarm/coalarm_service/CoalarmServiceApplication.java
@@ -1,7 +1,7 @@
 package _1danhebojo.coalarm.coalarm_service;
 
-import _1danhebojo.coalarm.coalarm_service.global.config.JwtProperties;
-import _1danhebojo.coalarm.coalarm_service.global.config.KakaoProperties;
+import _1danhebojo.coalarm.coalarm_service.global.properties.JwtProperties;
+import _1danhebojo.coalarm.coalarm_service.global.properties.KakaoProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
